move value into its own folder
